git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@4307 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/srcscintilla/iupsci_clipboard.c
@@ -50,7 +50,7 @@ static int iScintillaSetClipboardAttrib(Ihandle *ih, const char *value)
 
 static char* iScintillaGetCanPasteAttrib(Ihandle* ih)
 {
-  return iupStrReturnBoolean (IupScintillaSendMessage(ih, SCI_CANPASTE, 0, 0)); 
+  return iupStrReturnBoolean((int)IupScintillaSendMessage(ih, SCI_CANPASTE, 0, 0));
 }
 
 /***** Undo and Redo *****
@@ -77,7 +77,7 @@ static int iScintillaSetUndoAttrib(Ihandle *ih, const char *value)
 
 static char* iScintillaGetUndoAttrib(Ihandle* ih)
 {
-  return iupStrReturnBoolean (IupScintillaSendMessage(ih, SCI_CANUNDO, 0, 0)); 
+  return iupStrReturnBoolean((int)IupScintillaSendMessage(ih, SCI_CANUNDO, 0, 0));
 }
 
 static int iScintillaSetRedoAttrib(Ihandle *ih, const char *value)
@@ -91,7 +91,7 @@ static int iScintillaSetRedoAttrib(Ihandle *ih, const char *value)
 
 static char* iScintillaGetRedoAttrib(Ihandle* ih)
 {
-  return iupStrReturnBoolean (IupScintillaSendMessage(ih, SCI_CANREDO, 0, 0)); 
+  return iupStrReturnBoolean((int)IupScintillaSendMessage(ih, SCI_CANREDO, 0, 0));
 }
 
 static int iScintillaSetUndoCollectAttrib(Ihandle *ih, const char *value)
@@ -105,7 +105,7 @@ static int iScintillaSetUndoCollectAttrib(Ihandle *ih, const char *value)
 
 static char* iScintillaGetUndoCollectAttrib(Ihandle* ih)
 {
-  return iupStrReturnBoolean (IupScintillaSendMessage(ih, SCI_GETUNDOCOLLECTION, 0, 0)); 
+  return iupStrReturnBoolean((int)IupScintillaSendMessage(ih, SCI_GETUNDOCOLLECTION, 0, 0));
 }
 
 static int iScintillaSetUndoActionAttrib(Ihandle *ih, const char *value)
